Tic Tac Toe with reinforcement learning.
Creating an instance of this class will train (if episodes>0) a reinforcement learning algorithm to play tic tac toe, and launch a simple GUI so the user can play against the algorithm.

Setting episodes=0 means that the algorithm does not train, and it will make completely random moves. During every episode of training, the algorithm will literally play a game of tic tac toe against itself. It is rewarded for winning and penalized for losing, and it updates the value of observed states accordingly. I find that after around episodes=15000, the algorithm becomes unbeatable.

If the user plays optimally, every game will therefore end in a draw.

The GUI is very simple, as this was not the focus here.

Author: Jakob Lorenzen